api/pyproject.toml
[project]
name = "radiovier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "astral>=3.2",
    "datetime>=5.5",
    "fastapi>=0.115.8",
    "googletrans==4.0.0rc1",
    "janus>=2.0.0",
    "jinja2>=3.1.5",
    "psycopg2-binary>=2.9.10",
    "pydub>=0.25.1",
    "python-dateutil>=2.9.0.post0",
    "sse-starlette>=2.2.1",
    "uvicorn>=0.34.0",
    "vosk>=0.3.45",
]
